// src/app/clients/Client.ts
export interface IClient {
    accountId?: any;
    additionalInfo?: string;
    additionalPhone?: any;
    address: string;
    birthDay: number;
    birthMonth: number;
    birthYear: number;
    cardAvailability?: boolean;
    cardNumber?: number;
    contactPerson?: string;
    description: string;
    divisionNumber?: any;
    email: string;
    gender: string;
    howDidYouKnowAnother?: string;
    howDidYouKnowId?: any;
    id?: number;
    name: string;
    phone: string;
    tabNumber?: any;
    mapedBirthDay?: string;
}